requirements.txt
binance-connector
kafka-python
python-dotenv